REFACTOR Extracted react classes into individual files
