+++ src/App.jsx
@@ -1,108 +1,6 @@
 import React, { Component } from 'react';
 import { processInput, findMethod } from './suggestion.js';
-
-class History extends React.Component {
-  generateHistory() {
-    const items = this.props.history.map(item => item.slice());
-    return items.reverse().map((item, index) => <div key={index}>{item}</div>);
-  }
-
-  render() {
-    const list = this.generateHistory();
-    return (
-      <div>{list}</div>
-    );
-  }
-}
-
-class InputForm extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      input: null,
-      output: null,
-    };
-
-    this.handleChange = this.handleChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-  }
-
-  handleChange(event) {
-    const name = event.target.name;
-    this.setState({
-      [name]: event.target.value,
-    });
-  }
-
-  handleSubmit(event) {
-    event.preventDefault();
-    this.props.setInputOutput(this.state.input, this.state.output);
-  }
-
-  render() {
-    return (
-      <div>
-        <form onSubmit={this.handleSubmit}>
-          <input name="input" type="text" onChange={this.handleChange} />
-          <input name="output" type="text" onChange={this.handleChange} />
-          <input type="submit" value="Submit" />
-        </form>
-      </div>
-    );
-  }
-}
-
-class Summary extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      input: null,
-      output: null,
-      history: [],
-    };
-
-    this.setInputOutput = this.setInputOutput.bind(this);
-    this.logSuggestion = this.logSuggestion.bind(this);
-  }
-
-  setInputOutput(input, output) {
-    this.setState({
-      input,
-      output,
-    });
-  }
-
-  logSuggestion(input, output, code) {
-    this.state.history.push(`${input} > ${output} = ${code}`);
-  }
-
-  render() {
-    return (
-      <div>
-        <div>
-          User Input:
-          <InputForm
-            setInputOutput={this.setInputOutput}
-          />
-        </div>
-        <div>
-          Code Suggestion:
-          <CodeSuggestion
-            input={this.state.input}
-            output={this.state.output}
-            logSuggestion={this.logSuggestion}
-          />
-        </div>
-        <div>
-          History:
-          <History
-            history={this.state.history}
-          />
-        </div>
-      </div>
-    );
-  }
-}
+import { Summary } from './summary';
 
 class App extends React.Component {
   render() {
