+++ src/inputform.js
@@ -1,0 +1,38 @@
+import React, { Component } from 'react';
+import { setInputOutput } from './summary';
+export class InputForm extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      input: null,
+      output: null,
+    };
+
+    this.handleChange = this.handleChange.bind(this);
+    this.handleSubmit = this.handleSubmit.bind(this);
+  }
+
+  handleChange(event) {
+    const name = event.target.name;
+    this.setState({
+      [name]: event.target.value,
+    });
+  }
+
+  handleSubmit(event) {
+    event.preventDefault();
+    this.props.setInputOutput(this.state.input, this.state.output);
+  }
+
+  render() {
+    return (
+        <div>
+          <form onSubmit={this.handleSubmit}>
+            <input name="input" type="text" onChange={this.handleChange} />
+            <input name="output" type="text" onChange={this.handleChange} />
+            <input type="submit" value="Submit" />
+          </form>
+        </div>
+    );
+  }
+}
