+++ src/summary.js
@@ -1,58 +1,9 @@
 import React, { Component } from 'react';
-import { processInput, findMethod } from './suggestion.js';
+import { CodeSuggestion } from './suggestion.js';
+import { InputForm } from './inputform.js'
+import { History } from './history.js'
 
-class History extends React.Component {
-  generateHistory() {
-    const items = this.props.history.map(item => item.slice());
-    return items.reverse().map((item, index) => <div key={index}>{item}</div>);
-  }
-
-  render() {
-    const list = this.generateHistory();
-    return (
-        <div>{list}</div>
-    );
-  }
-}
-
-class InputForm extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      input: null,
-      output: null,
-    };
-
-    this.handleChange = this.handleChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-  }
-
-  handleChange(event) {
-    const name = event.target.name;
-    this.setState({
-      [name]: event.target.value,
-    });
-  }
-
-  handleSubmit(event) {
-    event.preventDefault();
-    this.props.setInputOutput(this.state.input, this.state.output);
-  }
-
-  render() {
-    return (
-        <div>
-          <form onSubmit={this.handleSubmit}>
-            <input name="input" type="text" onChange={this.handleChange} />
-            <input name="output" type="text" onChange={this.handleChange} />
-            <input type="submit" value="Submit" />
-          </form>
-        </div>
-    );
-  }
-}
-
-class Summary extends React.Component {
+export class Summary extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
